Get simple-binarise sort of working
Some JS/HTML changes to get it working with the image (rather than
asking the user to upload a file). Still have to add back the threshold
display and other things (and clean up the rest of the JS). Submitting
the form brings you to another view, which uses celery to delay a task
but it doesn't work yet