Add diva.js stuff
Complete with readme updates and a very basic test.

(There are a lot of failing tests, but this one should pass. I'll fix
the failing tests in another commit.)

Note that while IIPImage is required for diva.js to work, you don't
have to install it as a dependency or anything (it's installed on
Susato). If you do want to see what diva looks like, the easiest way
is to look at my development instance (rodan.simssa.ca:8001).