Fix long execution time for Rotate job with angle=0
the reason this task was taking so long to execute was because of the thumbnails that were being created. if the angle is 0, we can take the thumbnails of the previous job and use them for this one.

so i added a copy_thumbnails_from_previous_job() function inside utils.py and a new case under "tiff" output_type where we check output_content=="copy"

rotate now takes less than a second when angle==0, so i added a sleep(1s) inside the case where angle==0 otherwise the refresh (every 1s) on the project page wouldn't detect rotate task completion