Integrate barfinder algorithm into rodan
Currently uses a hardcoded staff group hint. Also uses a wrapper object
around the python binding to the MeiDocument since this is the format that
RODAN expects ... when a celery task returns MEI, it should either take
a python binding to an MEI document or a text version of the document.