WIP combine_mei project level job
should be in a workable state. added an extra decorator which is essentially a copy paste.. will need to work on it some more next week to make a cleaner solution, just wanted to get something in. also some hackiness here and there, again just wanted to make sure to have something workable for the weekend in case someone needs it