Major rewrite of the workflow system
This new system introduces two new models, a workflow run and a run job. Now when a workflow is run, the workflow jobs are copied to run jobs to get executed, and a workflow run is created to manage a single run of a workflow.

The saving of results has also changed. Now results are saved in a directory corresponding to the workflow run that created them (instead of with the page that they belong to).