Add Cappuccino wrappers for job interfaces
Barline-correction is the only one that I could not get working. All the rest are work somewhat. There is an issue with the on `window.onresize` which makes binarise and despeckle not work - I commented it out, brianstern should take a look at it. Also various other fixes inside the Javascript files to get the interfaces somewhat functional.