Add functionality to interactive crop view
Adds logic to differentiate GET and POST requests sent to the /interactive/crop/.

Logic for fetching the target run_job using a run_job's uuid field - This should be sent with the Cappuccino app.

<b>ex:</b> inside RKCrop.j, you can have something like
`[self setMainFrameURL:@"/interactive/crop?rj_uuid=6a1c1c4cfaf54e9fbf64473bcbeb81b3"];`

The code checks whether it is the first job in a workflow or a job somewhere down the line. It then gets the URL of the image that should be displayed in the interactive interface, and sends that off to the django template associated with the job.

The POST code is not complete - It checks whether a UUID has been provided, and then checks whether all the relevant parameters were sent with the POST request. If yes, it will change all the settings within the run_job object and set `needs_input` to false, which should allow the job to fall through and execute, within celery. This still needs to be implemented.