Updates non-interactive border_removal job, adds interactive border_removal job, and relocates the border_removal jobs inside "gamera/custom".
If you already have the old version of border_removal jobs, you will need to delete and recreate your database to see the changes, or manually delete all the border_removal jobs from the database and restart the server. Also, you need to have the latest version of Document Pre-processing Toolkit, and reinstall that.

Fixes #161.