Everything seems to work properly now except for drag and drop.  If you add a new poly or modify an existing one, drag/drop will not work correctly.
Having said that, is that functionality necessary?  Sounds like a "nice to have", but I imagine that 99% of the work will be adding points to polys...

I'm going to close this but add a new issue to take a look at the drag/drop.  We also may want to totally disable drag/drop for now.