Modified border removal and segmentation jobs to use the new base task class.
Also note that the settings now belong to the Task class instead of the module loader, because I felt that the settings should be as close as possible to the actual execution code. We need some robust settings/argument classes that will handle the settings so that we do not need to hardcode them. An extension of the gamera Args system will be great.

Refs #184. Fixes #189.