New: Sorting on job name in Job table.
Added a new method to the Job model that does the same as the transformer.  This allows the array controller to sort by the shortened name (was sorting by full name before).

Andrew, if there's a better way to do this (and revert back to the transformer), please let me know.