New: more specific uuid types.
The setting type uuid has been converted to uuid_classifier and uuid_pageglyphs for better type checking in the client. Although since it is easier for the client to send urls rather than uuids, this might soon be renamed to classifier_url and pageglyph_url. String parsing on the client side should be as minimal as possible.