Fix: type error in UUID parser.
The get_uuid_from_url function was raising "TypeError: expect string or buffer." My best guess was that this happened for unsupported unicode characters in the python regex function. Blindly casting the url to str seemed to have solved the problem, but it may have affected unicode url support in Rodan.