Tests: New tests can use fixtures
This commit adds two fixtures to Rodan which can be used to provide initial data for testing. This initial data is a dump of a mock Rodan project containing instances of Project, Page, Job, RunJob, WorkflowJob, Workflow, and Result. A second fixture contains two user records, one AnonymousUser, and one "ahankins" administrator user.

A sample test is contained in the views/test_project.py folder. The fixtures are specified at the beginning. When Django runs the tests it will first load the fixtures.