Fixed: Move initialization of the jobs system to URLs.
Previously, the jobs system was initialized when the Rodan __init__.py file was called. This caused problems if it was called prior to the DB being set up.

By moving it to the `urls.py` file we ensure that it is only called when the first request comes in. Perhaps not optimal, but we'll figure this out later.