Fixed: auth status now checks whether a token is passed
Previously the authentication status check would return false even if the request contained a valid Authentication token, since the auth check was not checking for it. This commit fixes this to check for an authentication token in the process.

It also changes the response of auth status to contain the user's token.