Fixed: improve parallel processing of workflowruns
Originally, master task finds all runable runjobs, runs them in a chord, and calls itself back after the chord is finished. Then the execution time always depends on the slowest runjob.

This commit fixes this issue by calling master task after finishing every runjob. However, it needs a lock on the runjob, to avoid multiple master tasks assigning the same runjob. The lock field stores the thread ID of master task when the runjob is locked. And this field is not designed to be provided to the client.