Maintenance: remove session authentication codes
We decide not to use session auth as it hinders the scalability of RESTful APIs.
Rodan clients will need to use token auth. But in DEBUG mode, we need REST framework
browsable API and Django admin interface, which are old-style HTTP pages.

The solution is to use HTTP basic auth. It is enabled only in DEBUG mode as it is
not that safe compared to token auth.