Modified: Input/output port types. See description [WIP]
The input and output ports on a job are very volatile during development. For this reason, the ability to change the name of an input/output port has been added to their models. Jobs will remain in the workflows, but a check should be added to  workflowjobs when `input_port_types` and `output_port_types` are removed. (It should repopulate the correct value automatically)