Modified: Restructured settings for job queues
The rodan core image must have ALL rodan jobs listed in settings.py, because it loads all the jobs into the database.

Python2 jobs can not be in the python3 list because it will try to load them anyway and crash. This is only important when working as a volume mount of the rodan folder into the docker containers. When the containers are up, it does not matter very much because they each have their own settings.py anyway (different containers, different fs).