Don't show description of workflowrun in list
Description field has no size limit and can contain a list of
information for the entire workflow run. Depending on the job, it may
contain a large amount of info as it will serialize the settings object
with can contain any information needed to be preserved in phases of an
interactive job. Since this can be so big and massively increase the
response object size and is also not very helpful for finding a
workflowrun, remove it.