Fix unseen variable error assigning the input port type to list.
This functionality will not be trigerred in rodan v2.0.0 since is_list is
not working, but this commit still fix some bugs.

When setting is_list=True, rodan.jobs.base.py will try to read available resources
assigned to the port. This is done by using os.listdir and os.path.isfile with list
comprehension. os.path.isfile tries to use the f variable which does not exist. Change
it to ff.