Fix authentication issue when accessing Django admin page
Updated authentication handling in the Django admin page to address an issue introduced in Django 2.0.13. The issue was related to the django.contrib.auth.authenticate method expecting string instead of byte values for usernames and passwords. This fix converts byte values to strings to align with the change in Django version.